[Bronze V] Title: 이상한 기호, Time: 128 ms, Memory: 14180 KB -BaekjoonHub

+++ 백준/Bronze/15964. 이상한 기호/이상한 기호.java
@@ -13,6 +13,6 @@ public class Main {
         int A = Integer.parseInt(st.nextToken());
         int B = Integer.parseInt(st.nextToken());
 
-        System.out.print((A+B)*(A-B));
+        System.out.print((long) (A+B)*(A-B));
     }
 }
